pyproject.toml
[tool.poetry]
name = "bet-maker"
version = "0.1.0"
description = ""
authors = ["as.petryanin <aspetryanin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
SQLAlchemy = "^2.0.28"
ruff = "^0.3.2"
black = "^24.2.0"
pydantic-settings = "^2.2.1"
colorlog = "^6.8.2"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
reorder-python-imports = "^3.12.0"
pre-commit = "^3.6.2"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
select = ["D", "E", "F", "W", "N", "C", "B", "RUF100"]
ignore = [
    "B008",
    "D200",
    "D202",
    "D203",
    "D205",
    "D210",
    "D212",
    "D213",
    "D400",
    "D406",
    "D407",
    "D413",
    "D415",
    "N815",
]
exclude = [
    "versions",
]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests",
]
asyncio_mode="auto"
